Update bit-read function
Seems to be working now. Work left is to convert the read outs to a
value to be returned.

Thread/sleep was replaced with a specific sleep function. The sleep
function is a bit punishing for the CPU, but only sleeps for 100
nanoseconds. Given that we have dual core on the Zedboard, this should
be a small price to pay to reduce the time read/writing to the analog FPGA.